[util] Extract a utility function for killing scripts across all hosts
